requirements.txt
absl-py==0.13.0
argon2-cffi==20.1.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
datasets==1.11.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.7.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
huggingface-hub==0.0.12
idna==3.2
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
multiprocess==0.70.12.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
rouge==1.0.1
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
tokenizers==0.10.3
tornado==6.1
tqdm==4.62.1
traitlets==5.0.5
transformers==4.9.2
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xxhash==2.0.2
